+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.0
+  * Bump dependency versions for singer-python and request. [#17](https://github.com/singer-io/tap-mailshake/pull/17)
+  * Fixed schema mismatch issues, updated schema for all existing streams.[#18](https://github.com/singer-io/tap-mailshake/pull/18)
+  * Changed replication method for team_members INCREMENTAL to FULL_TABLE and fixed sync issues.[#18](https://github.com/singer-io/tap-mailshake/pull/18)
+
 ## 0.1.0
   * Bump dependency versions for twistlock compliance [#16](https://github.com/singer-io/tap-mailshake/pull/16)